chore: restructure Kafka codebase
