+++ frontend/src/components/MessagingQueueHealthCheck/AttributeCheckList.tsx
@@ -17,15 +17,14 @@ import { QueryParams } from 'constants/query';
 import ROUTES from 'constants/routes';
 import { History } from 'history';
 import { Bolt, Check, OctagonAlert, X } from 'lucide-react';
+import {
+	KAFKA_SETUP_DOC_LINK,
+	MessagingQueueHealthCheckService,
+} from 'pages/MessagingQueues/MessagingQueuesUtils';
 import { ReactNode, useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { isCloudUser } from 'utils/app';
 import { v4 as uuid } from 'uuid';
-
-import {
-	KAFKA_SETUP_DOC_LINK,
-	MessagingQueueHealthCheckService,
-} from '../MessagingQueuesUtils';
 
 interface AttributeCheckListProps {
 	visible: boolean;
