+++ frontend/src/components/MessagingQueueHealthCheck/MessagingQueueHealthCheck.tsx
@@ -5,9 +5,9 @@ import { Button } from 'antd';
 import cx from 'classnames';
 import { useOnboardingStatus } from 'hooks/messagingQueue/useOnboardingStatus';
 import { Bolt, FolderTree } from 'lucide-react';
+import { MessagingQueueHealthCheckService } from 'pages/MessagingQueues/MessagingQueuesUtils';
 import { useEffect, useMemo, useState } from 'react';
 
-import { MessagingQueueHealthCheckService } from '../MessagingQueuesUtils';
 import AttributeCheckList from './AttributeCheckList';
 
 interface MessagingQueueHealthCheckProps {
