+++ frontend/src/components/MessagingQueues/MQCommon/MQCommon.styles.scss
